setup(UserContext): Dedicated types folder/file for context
